requirements.txt
flask==3.0.0
playwright==1.40.0
pypdf2==3.0.1
pillow==10.1.0
pytesseract==0.3.10
python-dotenv==1.0.0
sqlalchemy==2.0.23
imap-tools==1.5.0
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.3
pydantic[email]==2.5.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
flask-cors==4.0.0
flask-limiter==3.5.0
